+++ server/src/include/object.h
@@ -32,7 +32,7 @@
 
 #define decrease_ob(xyz) decrease_ob_nr(xyz,1)
 
-#define WEIGHT(op) ((op->nrof ? op->weight * op->nrof : op->weight) + op->carrying)
+#define WEIGHT(op) ((op->nrof ? op->weight * op->nrof : (unsigned int) (op->weight) + op->carrying))
 #define WEIGHT_NROF(op, nrof) ((nrof ? op->weight * nrof : op->weight) + op->carrying)
 
 /* move_apply() function call flags */
@@ -190,9 +190,9 @@ typedef struct obj
     uint8           item_race;          /* item crafted from race x. "orcish xxx", "dwarven xxxx" */
     uint8           item_level;         /* level needed to use or apply this item */
     uint8           item_skill;         /* if set and item_level, item_level in this skill is needed */
-    
+
     sint8           anim_enemy_dir;     /**< special shadow variable: show dir to targeted enemy
-                                          for mobs: activate attack animation 
+                                          for mobs: activate attack animation
                                           for the given direction unless == -1 */
     sint8           anim_moving_dir;    /* sic: shows moving dir or -1 when object do something else */
